Initial migration from ava to jest
