+++ shell.nix
@@ -22,8 +22,8 @@ in
 
       export PATH="$(pwd)/dist/bin:$(npm bin):$PATH"
 
-      # setting up non-js dev dependencies
-      npm install awesome-typescript-loader source-map-loader
+      # Install jest cli globally
+      npm install -g jest-cli
 
       # setting up for nix-build
       npm install --package-lock-only
