+++ tsconfig.json
@@ -12,7 +12,6 @@
       "@matrixai/*": ["*"]
     }
   },
-  "include": [
-    "./src/**/*"
-  ]
+  "include": [ "./src/**/*" ],
+  "exclude": [ "**/*.test.ts" ]
 }
